requirements.txt
Appium-Python-Client==0.48
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.1
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pytest==5.2.2
selenium==3.141.0
six==1.12.0
urllib3==1.25.6
wcwidth==0.1.7
zipp==0.6.0
